+++ NRI/Source/D3D12/TextureD3D12.h
@@ -78,7 +78,7 @@ namespace nri
         if (dimension == 2)
             size = (uint16_t)m_TextureDesc.DepthOrArraySize;
 
-        size = GetAlignedSize( std::max(size >> mipOffset, 1), GetTexelBlockWidth(m_Format) );
+        size = (uint16_t)GetAlignedSize( std::max(size >> mipOffset, 1), GetTexelBlockWidth(m_Format) );
 
         const uint32_t resourceDimension = m_TextureDesc.Dimension - D3D12_RESOURCE_DIMENSION_TEXTURE1D;
         return (dimension <= resourceDimension) ? size : 1;
